AtS E3S9: fix drawing glitch when filling the map with Xv^Gov
We were resetting the screen RGB adjustment to 0,0,0 *before* committing
terrain changes, resulting in the previous map being temporarily visible
for a few milliseconds (more specifically, for as long as it takes for
Wesnoth to rebuild the gamemap terrain graphics).

This commit changes the sequence order so that terrain changes are
committed *before* resetting the screen RGB adjustment to 0,0,0.