AtS E2S4: try to minimize animation glitches during the initial cutscene
Namely, the kind of glitch where a unit comes into view and gets drawn
before we get a chance to play their post_teleport animation. This is
pretty much inevitable, but by changing the order of the WML actions
affecting Elorran, and spawning Tara at the target location pre-hidden,
the effect seems to disappear on a high-end machine.