AtS: workaround [lock_view] + [move_unit_fake]/[move_unit] interaction bug on Wesnoth 1.11.x
Due to an oversight of mine when implementing the lock_view/unlock_view
feature, unit movement is affected by the lock_view option as though the
"Follow unit actions" was disabled in Advanced Preferences.

In order to allow [move_unit_fake] and [move_unit] (which is implemented
using [move_unit_fake]) to scroll the viewport as necessary when the
lock_view option is enabled, we temporarily unset that option before
passing control back to the mainline (C++) implementation of the WML
action.

This should hopefully be fixed in a future development release when I
figure out the best way to fix it C++-ways.